requirements.txt
google-generativeai
python-dotenv
streamlit
pymupdf
transformers
langchain
